Fix region selection
- Fix incorrect replacement of request.region in the region selection template
- Fix missing redirection in case an admin user opens the region selection
