+++ integreat_cms/cms/views/dashboard/__init__.py
@@ -4,4 +4,3 @@ This package contains all dashboard views
 from .admin_dashboard_view import AdminDashboardView
 from .dashboard_view import DashboardView
 from .region_selection import RegionSelection
-from .redirect_view import RedirectView
